Support decimal prices; align rental table to production schema names
Prices lost decimals because the model declared member_price/non_member_price
as Integer. Restore the legacy behavior (old system used DOUBLE) and match the
existing production schema, which the port had "corrected":
- table rental_template -> rental_tamplate (Python class/attrs unchanged)
- member_price/non_member_price Integer -> Float (DOUBLE)
- DB column minutes_offset -> minuts_offset (attr stays minutes_offset)
- available_courts_search FK rental_template_id -> rental_tamplate_id
- pydantic int -> float (MatrixCell, CourtSlotOut, effective_price)
- schedule editor: step=0.1 on price inputs + round to 1 decimal

Existing new-schema DBs need a one-off ALTER (rename table/columns, prices to
DOUBLE); production already uses these names.

+++ backend/app/routers/courts.py
@@ -24,9 +24,9 @@ class CourtSlotOut(BaseModel):
     date: date
     hour: int
     minutes_offset: int
-    member_price: int | None
-    non_member_price: int | None
-    price: int          # effective price for the requesting user (member_price for members)
+    member_price: float | None
+    non_member_price: float | None
+    price: float        # effective price for the requesting user (member_price for members)
     is_member_price: bool
     is_free: bool       # member_price == 0 → no payment needed
     covered_by_subscription: bool = False   # user holds a מנוי → books free via subscription
